+++ todo.md
@@ -134,3 +134,4 @@ println ! ("Type: {:?} Data: {:?}", event_type, data);
 - [x] Finish the connect implementation
 - [x] Implement ws mock for unit test ...
 - [x] Fix print to logger messages
+    [src/lib/replication_ops.rs:128 ]
